rdrr.io Find an R package R language docs Run R in your browser

biosampleR
Biodiversity Index Calculation and Bootstrap Confidence Interval Estimation

Package index
Vignettes

README.md

biosampleR-example_workflow

Functions
11

Source code
13

Man pages
6

BCI: Barro-Colorado Island Tree Counts

calc_delta_var: Calculate the change in variance with increasing number of...

calc_diversity_indices: Calculate biodiversity summary indices from count data

create_resample: Create multiple resamples of a data set.

generate_subsamples: Generate subsamples of a data frame with a number of sites...

get_sample_stats: Calculate biodiversity measures and summary statistics for a...

Browse all...

Home
/
CRAN
/
biosampleR
/
calc_diversity_indices: Calculate biodiversity summary indices from count data

calc_diversity_indices: Calculate biodiversity summary indices from count data
In biosampleR: Biodiversity Index Calculation and Bootstrap Confidence Interval Estimation

View source: R/calc_diversity_indices.R

calc_diversity_indices	R Documentation

Calculate biodiversity summary indices from count data

Description

Calculate biodiversity summary indices from count data

Usage

calc_diversity_indices(data)

Arguments

data
A data frame of count data, with sites as rows and species as columns.

Value

A data frame with sites as rows and diversity indices as columns. The columns are: abundance, species richness, Shannon diveristy index, Simpson diversity index, Chao1, Difference between Choa1 and species richness.

Examples

ind <- calc_diversity_indices(BCI)

biosampleR documentation built on Sept. 14, 2023, 1:07 a.m.

Related to calc_diversity_indices in biosampleR...

biosampleR index

README.md biosampleR-example_workflow

What can we improve?

Extra info (optional)