rdrr.io Find an R package R language docs Run R in your browser

iNEXT
Interpolation and Extrapolation for Species Diversity

Package index
Vignettes

Package overview

Introduction Aug. 2022

Functions
64

Source code
10

Man pages
17

ant: Ant data (datatype = "incidence_freq")

as.abucount: Transform abundance raw data to abundance row-sum counts...

as.incfreq: Transform incidence raw data to incidence frequencies (iNEXT...

bird: Bird data (datatype = "abundance")

ChaoRichness: Estimation of species richness

ChaoShannon: Estimation of Shannon entropy/diversity

ChaoSimpson: Estimation of Gini-Simpson index or Simpson diversity

ciliates: Ciliates data (datatype = "incidence_raw")

DataInfo: Exhibit basic data information

estimateD: Compute species diversity with a particular level of sample...

fortify.iNEXT: Fortify method for classes from the iNEXT package.

ggiNEXT: ggplot2 extension for an iNEXT object

iNEXT: iNterpolation and EXTrapolation of Hill numbers

iNEXT-package: Interpolation and extrapolation for species diversity

plot.iNEXT: Plotting iNEXT object

print.iNEXT: Printing iNEXT object

spider: Spider data (datatype = "abundance")

Browse all...

Home
/
CRAN
/
iNEXT
/
spider: Spider data (datatype = "abundance")

spider: Spider data (datatype = "abundance")
In iNEXT: Interpolation and Extrapolation for Species Diversity

spider	R Documentation

Spider data (datatype = "abundance")

Description

The data include spider species abundances in two canopy manipulation treatments (Girdled and Logged) of hemlock trees (Ellison et al. 2010, Sackett et al. 2011). In the Girdled treatment site, there were 26 species among 168 individuals; in the Logged treatment site, there were 37 species among 252 individuals.

Usage

data(spider)

Format

The format for each site is a list of species abundances:
A list of 2 vectors
$ Girdled: num [1:26] 46 22 17 15 15 9 8 6 6 4 ...
$ Logged : num [1:37] 88 22 16 15 13 10 8 8 7 7 ...

References

Ellison, A.M., Barker-Plotkin, A.A., Foster, D.R. & Orwig, D.A. (2010) Experimentally testing the role of foundation species in forests: the Harvard forest hemlock removal experiment. Methods in Ecology and Evolution, 1, 168-179.

Sackett, T.E., Record, S., Bewick, S., Baiser, B., Sanders, N.J. & Ellison, A.M. (2011) Response of macroarthropod assemblages to the loss of hemlock (Tsuga canadensis), a foundation species. Ecosphere, 2, art74.

iNEXT documentation built on Aug. 8, 2025, 6:41 p.m.

Related to spider in iNEXT...

iNEXT index

Package overview Introduction Aug. 2022

What can we improve?

Extra info (optional)