rdrr.io Find an R package R language docs Run R in your browser

vegan
Community Ecology Package

Package index
Vignettes

Package overview

Design decisions and implementation

Diversity analysis in vegan

Introduction to ordination in vegan

Partition of Variation

Functions
956

Source code
444

Man pages
112

add1.cca: Add or Drop Single Terms to a Constrained Ordination Model

adipart: Additive Diversity Partitioning and Hierarchical Null Model...

adonis: Permutational Multivariate Analysis of Variance Using...

anosim: Analysis of Similarities

anova.cca: Permutation Test for Constrained Correspondence Analysis,...

avgdist: Averaged Subsampled Dissimilarity Matrices

BCI: Barro Colorado Island Tree Counts

beals: Beals Smoothing and Degree of Absence

betadisper: Multivariate homogeneity of groups dispersions (variances)

betadiver: Indices of beta Diversity

bgdispersal: Coefficients of Biogeographical Dispersal Direction

bioenv: Best Subset of Environmental Variables with Maximum (Rank)...

biplot.rda: PCA biplot

cascadeKM: K-means partitioning using a range of values of K

cca: [Partial] [Constrained] Correspondence Analysis and...

cca.object: Result Object from Constrained Ordination

CCorA: Canonical Correlation Analysis

clamtest: Multinomial Species Classification Method (CLAM)

commsim: Create an Object for Null Model Algorithms

contribdiv: Contribution Diversity Approach

dbrda: Principal Coordinates Analysis and [Partial] Distance-based...

decorana: Detrended Correspondence Analysis and Basic Reciprocal...

decostand: Standardization Methods for Community Ecology

designdist: Design your own Dissimilarities

deviance.cca: Statistics Resembling Deviance and AIC for Constrained...

dispindmorisita: Morisita index of intraspecific aggregation

dispweight: Dispersion-based weighting of species counts

distconnected: Connectedness of Dissimilarities

diversity: Ecological Diversity Indices

dune: Vegetation and Environment in Dutch Dune Meadows.

dune.taxon: Taxonomic Classification and Phylogeny of Dune Meadow Species

eigenvals: Extract Eigenvalues from an Ordination Object

envfit: Fits an Environmental Vector or Factor onto an Ordination

eventstar: Scale Parameter at the Minimum of the Tsallis Evenness...

fisherfit: Fit Fisher's Logseries and Preston's Lognormal Model to...

goodness.cca: Diagnostic Tools for [Constrained] Ordination (CCA, RDA, DCA,...

goodness.metaMDS: Goodness of Fit and Shepard Plot for Nonmetric...

indpower: Indicator Power of Species

influence.cca: Linear Model Diagnostics for Constrained Ordination

isomap: Isometric Feature Mapping Ordination

kendall.global: Kendall coefficient of concordance

linestack: Plots One-dimensional Diagrams without Overwriting Labels

make.cepnames: Abbreviates a Two-Part Botanical or Zoological Latin Name...

mantel: Mantel and Partial Mantel Tests for Dissimilarity Matrices

mantel.correlog: Mantel Correlogram

MDSaddpoints: Add New Points to NMDS ordination

MDSrotate: Rotate First MDS Dimension Parallel to an External Variable

metaMDS: Nonmetric Multidimensional Scaling with Stable Solution from...

mite: Oribatid Mite Data with Explanatory Variables

monoMDS: Global and Local Non-metric Multidimensional Scaling and...

Browse all...

Home
/
CRAN
/
vegan
/
dune: Vegetation and Environment in Dutch Dune Meadows.

dune: Vegetation and Environment in Dutch Dune Meadows.
In vegan: Community Ecology Package

dune	R Documentation

Vegetation and Environment in Dutch Dune Meadows.

Description

The dune meadow vegetation data, dune, has cover class values of 30 species on 20 sites. The corresponding environmental data frame dune.env has following entries:

Usage

  data(dune)
  data(dune.env)

Format

dune is a data frame of observations of 30 species at 20 sites. The species names are abbreviated to 4+4 letters (see make.cepnames). The following names are changed from the original source (Jongman et al. 1987): Leontodon autumnalis to Scorzoneroides, and Potentilla palustris to Comarum.

dune.env is a data frame of 20 observations on the following 5 variables:

A1:

a numeric vector of thickness of soil A1 horizon.

Moisture:

an ordered factor with levels: 1 < 2 < 4 < 5.

Management:

a factor with levels: BF (Biological farming), HF (Hobby farming), NM (Nature Conservation Management), and SF (Standard Farming).

Use:

an ordered factor of land-use with levels: Hayfield < Haypastu < Pasture.

Manure:

an ordered factor with levels: 0 < 1 < 2 < 3 < 4.

Source

Jongman, R.H.G, ter Braak, C.J.F & van Tongeren, O.F.R. (1987). Data Analysis in Community and Landscape Ecology. Pudoc, Wageningen.

Examples

data(dune)
data(dune.env)

vegan documentation built on May 25, 2026, 5:10 p.m.

Related to dune in vegan...

vegan index

Package overview Design decisions and implementation Diversity analysis in vegan Introduction to ordination in vegan Partition of Variation

What can we improve?

Extra info (optional)